add hash routing
