+++ apps/desktop/package.json
@@ -4,13 +4,13 @@
   "version": "1.0.0",
   "description": "My Electron application description",
   "main": ".webpack/main",
-	"proxy": "http://localhost:4000",
+  "proxy": "http://localhost:4000",
   "scripts": {
     "start": "electron-forge start",
     "package": "electron-forge package",
     "make": "electron-forge make",
     "publish": "electron-forge publish",
-		"clean": "rm prevVideos/*"
+    "clean": "rm prevVideos/*"
   },
   "devDependencies": {
     "@electron-forge/cli": "^7.6.0",
@@ -43,6 +43,7 @@
     "electron-squirrel-startup": "^1.0.1",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.2.0",
     "video.js": "^8.21.0"
   }
 }
