+++ apps/desktop/src/pages/SharedPage.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const SharedPage = () => {
+    return (
+        <div>
+            <h2>Shared Clips (Coming Soon)</h2>
+        </div>
+    );
+};
+
+export default SharedPage;
